pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "RAG chatbot for Typeform Help Center"
readme = "README.md"
authors = [{ name = "Spyroula Masiala" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.109.0",
  "uvicorn[standard]==0.27.0",
  "pydantic==2.5.3",
  "pydantic-settings==2.1.0",
  "openai==1.10.0",
  "langchain==0.1.4",
  "langchain-openai==0.0.5",
  "langchain-community==0.0.16",
  "pinecone-client==3.0.2",
  "python-dotenv==1.0.0",
  "python-multipart==0.0.6",
  "structlog==24.1.0"
]

[dependency-groups]
dev = [
  "pytest==7.4.4",
  "pytest-asyncio==0.23.3",
  "httpx==0.26.0"
]
